requirements.txt
streamlit==1.40.2
google-generativeai==0.8.3
edge-tts==7.2.8
moviepy==1.0.3
youtube-transcript-api==0.6.2
openai-whisper==20240930
pydub==0.25.1
python-dotenv==1.0.1
requests==2.32.3
Pillow==11.0.0
scipy==1.14.1
